Fixed memory test

+++ src/m68040/bios_m68040.c
@@ -211,7 +211,7 @@ unsigned long syscall_dispatch(int32_t function, int32_t param0, int32_t param1,
                     return proc_run((char *)param0, (int)param1, (char **)param2);
 
                 case KFN_MEM_GET_RAMTOP:
-                    return mem_get_ramtop();
+                    return (unsigned long)mem_get_ramtop();
 
                 case KFN_MEM_RESERVE:
                     return mem_reserve((unsigned long)param0);
